# Girvan newman optimization