### Summary

This PR optimizes the `girvan_newman` community detection algorithm in NetworkX.

The Girvan-Newman method is known to be computationally expensive, and users have reported performance concerns:
- [StackOverflow: Why is the Girvan-Newman algorithm in NetworkX so slow?](https://stackoverflow.com/questions/62951320/why-is-the-girvan-newman-algorithm-in-networkx-so-slow)
- [Issue #7992: Improve runtime of `girvan_newman()` function](https://github.com/networkx/networkx/issues/7992)

In [#7992](https://github.com/networkx/networkx/issues/7992), two optimization opportunities were identified:
1. **Cache the number of edges** and update it incrementally after every edge removal.
2. **Operate on one graph per connected component** to allow better caching of edge betweenness centrality.

This PR implements the **first optimization**: the number of edges is now cached and updated incrementally.  
The **second optimization** is not implemented here because it would require a deeper refactoring: the current method accepts a `most_valuable_edge` function that operates on the **full graph**, making per-component caching of betweenness centrality non-trivial.

### Performance

Benchmarks comparing current and optimized versions (averaged over 100 runs):

| Graph Type                   | Current Time/Call (s) | Optimized Time/Call (s) | Speedup |
|-------------------------------|-----------------------|-------------------------|---------|
| Complete Graph (n=10)         | 0.016050               | 0.015241                 | 1.05×   |
| Complete Graph (n=20)         | 0.170338               | 0.149063                 | 1.14×   |
| G(n=50, p=0.1)                | 1.068046               | 1.092481                 | 0.98×   |
| G(n=50, p=0.2)                | 4.944922               | 4.612656                 | 1.07×   |


Performance improvements are clear for larger or denser graphs, and small changes for easier graphs are inconclusive and likely due to normal benchmark variability.
Speedups range between ~0–14% depending on the graph structure.

### Benchmarking Code

The following code was used to measure runtimes:

```python
import timeit

setup = "import networkx as nx"

benchmarks = [
    ("Complete Graph (n=10)", "list(nx.community.girvan_newman(nx.complete_graph(10)))", 1000),
    ("Complete Graph (n=20)", "list(nx.community.girvan_newman(nx.complete_graph(20)))", 100),
    ("G(n=50, p=0.1)", "list(nx.community.girvan_newman(nx.fast_gnp_random_graph(50, p=0.2, seed=1)))", 100),
    ("G(n=50, p=0.2)", "list(nx.community.girvan_newman(nx.fast_gnp_random_graph(50, p=0.5, seed=1)))", 100),
]

for label, stmt, c in benchmarks:
    duration = timeit.timeit(stmt, setup=setup, number=c)
    per_call = duration / c
    print(f"{label:<30} | Total time: {duration:.4f}s | Time per call: {per_call:.6f}s")