@amcandio 2025-04-30
Just realized my benchmarking code is wrong bc it includes the graph generation time 😅