@dschult 2025-04-30
Another feature of benchmark time that might be controversial  idk...
It might be more accurate timing to take the minimum time of many runs instead of the average time. To the extent that other system activity slow down a run, using the minimum could remove the effects of system activity on the benchmarks. :}/