@amcandio 2025-05-24
I doubt this optimization will be significant so let's focus on https://github.com/networkx/networkx/pull/8000 for now